fixed export and first step
