+++ src/core/onResize.ts
@@ -1,5 +1,5 @@
 import refresh from "./refresh";
-import { IntroJs } from "../index";
+import { IntroJs } from "../IntroJs";
 
 export default function onResize(this: IntroJs) {
   refresh.call(this);
