+++ src/core/showElement.ts
@@ -12,7 +12,7 @@ import removeShowElement from "./removeShowElement";
 import createElement from "../util/createElement";
 import setStyle from "../util/setStyle";
 import appendChild from "../util/appendChild";
-import { IntroItem, IntroJs } from "../index";
+import { IntroItem, IntroJs } from "../IntroJs";
 
 /**
  * Gets the current progress percentage
